+++ src/deepqmc/wf/paulinet/paulinet.py
@@ -17,7 +17,7 @@ from .omni import OmniSchNet, SubnetFactory
 from .pyscfext import pyscf_from_mol
 from .schnet import ElectronicSchNet
 
-__version__ = '0.1.0'
+__version__ = '0.1.1'
 __all__ = ['PauliNet']
 
 log = logging.getLogger(__name__)
@@ -173,14 +173,12 @@ class PauliNet(WaveFunction):
             if backflow_factory
             else None
         )
-        self.r_backflow = None
         if omni_factory:
             assert not backflow_factory and not jastrow_factory
             self.omni = omni_factory(
                 mol, dist_feat_dim, n_up, n_down, *backflow_spec, **(omni_kwargs or {})
             )
             self.backflow = self.omni.forward_backflow
-            self.r_backflow = self.omni.forward_r_backflow
             self.jastrow = self.omni.forward_jastrow
         else:
             self.omni = None
@@ -362,16 +360,11 @@ class PauliNet(WaveFunction):
             if self.jastrow or self.mo.net
             else None
         )
-        if self.r_backflow or self.backflow or self.cusp_same or self.jastrow:
+        if self.backflow or self.cusp_same or self.jastrow:
             dists_elec = pairwise_distance(rs, rs)
-        if self.r_backflow or self.backflow or self.jastrow:
+        if self.backflow or self.jastrow:
             edges_nuc = edges_nuc[n_atoms:].view(batch_dim, n_elec, n_atoms, -1)
             edges = self.dist_basis(dists_elec), edges_nuc
-        if self.r_backflow:
-            rs_flowed = self.r_backflow(rs, *edges)
-            diffs_nuc = pairwise_diffs(
-                torch.cat([coords, rs_flowed.flatten(end_dim=1)]), coords
-            )
         xs = self.mo(diffs_nuc)
         # get orbitals as [bs, 1, i, mu]
         xs = xs.view(batch_dim, 1, n_elec, -1)