+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `PauliNet`:
     - `mo_factory`, functionality replaced with a mean-field backflow
+    - Real-space backflow
 
 ## [0.2.0] - 2020-08-19